fixed issue with the composition displaying incorrectly and added semicolon detection
