+++ .gitignore
@@ -2,4 +2,6 @@
 .project
 out/
 /bin/
+src/com/carlosjacob/exceptions/
 LexerOutput.txt
+
