--- src/Main.java
@@ -1,6 +1,0 @@
-public class Main {
-
-    public static void main(String[] args) {
-        Syntax syntax = new Syntax(args[0]);
-    }
-}
