+++ src/com/carlosjacob/Main.java
@@ -1,0 +1,8 @@
+package com.carlosjacob;
+
+public class Main {
+
+    public static void main(String[] args) {
+        Syntax syntax = new Syntax(args[0]);
+    }
+}
